explicit instantiation - FluidsQuasiImplicit
